# Optimize label comparisons and remove restrictions on types that can be labels