--- a/crates/bevy_ecs/src/schedule/label.rs
+++ b/crates/bevy_ecs/src/schedule/label.rs
@@ -34,7 +34,7 @@ define_label!(
 // Implement string-labels for now.
 
 #[doc(hidden)]
-pub static STR_INTERN: Labels = Labels::new();
+pub static STR_INTERN: TypedLabels<&'static str> = TypedLabels::new();
 
 /// Implements a label trait for `&'static str`, the string literal type.
 #[macro_export]
@@ -42,7 +42,7 @@ macro_rules! impl_string_label {
     ($label:ident) => {
         impl $label for &'static str {
             fn data(&self) -> u64 {
-                $crate::schedule::STR_INTERN.intern::<Self>(self)
+                $crate::schedule::STR_INTERN.intern(self)
             }
             fn fmt(idx: u64, f: &mut std::fmt::Formatter) -> std::fmt::Result {
                 $crate::schedule::STR_INTERN
@@ -58,6 +58,59 @@ impl_string_label!(StageLabel);
 impl_string_label!(AmbiguitySetLabel);
 impl_string_label!(RunCriteriaLabel);
 
+/// A data structure used to intern a set of labels of a single concrete type.
+/// To store multiple distinct types, or generic types, try [`Labels`].
+pub struct TypedLabels<T: Clone + Hash + Eq>(
+    // The `IndexSet` is a hash set that preservers ordering as long as
+    // you don't remove items (which we don't).
+    // This allows us to have O(~1) hashing and map each entry to a stable index.
+    RwLock<IndexSet<T>>,
+);
+
+/// The type returned from [`Labels::get`](Labels#method.get).
+///
+/// Will hold a lock on the label interner until this value gets dropped.
+pub type LabelGuard<'a, L> = parking_lot::MappedRwLockReadGuard<'a, L>;
+
+impl<T: Clone + Hash + Eq> TypedLabels<T> {
+    pub const fn new() -> Self {
+        Self(RwLock::new(IndexSet::with_hasher(bevy_utils::FixedState)))
+    }
+
+    /// Interns a value, if it was not already interned in this set.
+    ///
+    /// Returns an integer used to refer to the value later on.
+    pub fn intern(&self, val: &T) -> u64 {
+        use parking_lot::RwLockUpgradableReadGuard as Guard;
+
+        // Acquire an upgradeable read lock, since we might not have to do any writing.
+        let set = self.0.upgradable_read();
+
+        // If the value is already interned, return its index.
+        if let Some(idx) = set.get_index_of(val) {
+            return idx as u64;
+        }
+
+        // Upgrade to a mutable lock.
+        let mut set = Guard::upgrade(set);
+        let (idx, _) = set.insert_full(val.clone());
+        idx as u64
+    }
+
+    /// Allows one to peek at an interned label and execute code,
+    /// optionally returning a value.
+    ///
+    /// Returns `None` if there is no interned label with that key.
+    pub fn scope<U>(&self, idx: u64, f: impl FnOnce(&T) -> U) -> Option<U> {
+        self.0.read().get_index(idx as usize).map(f)
+    }
+
+    /// Gets a reference to the label with specified index.
+    pub fn get(&self, idx: u64) -> Option<LabelGuard<T>> {
+        RwLockReadGuard::try_map(self.0.read(), |set| set.get_index(idx as usize)).ok()
+    }
+}
+
 /// Data structure used to intern a set of labels.
 ///
 /// To reduce lock contention, each kind of label should have its own global instance of this type.
@@ -65,19 +118,12 @@ impl_string_label!(RunCriteriaLabel);
 /// For generic labels, all labels with the same type constructor must go in the same instance,
 /// since Rust does not allow associated statics. To deal with this, specific label types are
 /// specified on the *methods*, not the type.
+///
+/// If you need to store a single concrete type, [`TypedLabels`] is more efficient.
 pub struct Labels(
-    // This type-map stores instances of `IndexSet<T>`, for any fully-resolved type `T`.
-    //
-    // The `IndexSet` is what interns the values for each type -- it's a hash set
-    // that preservers ordering as long as you don't remove items (which we don't).
-    // This allows us to use O(~1) hashing and map each entry to a stable index.
+    // This type-map stores instances of `IndexSet<T>`, for any `T`.
     RwLock<TypeMap>,
 );
-
-/// The type returned from [`Labels::get`](Labels#method.get).
-///
-/// Will hold a lock on the string interner for type `L`, until this value gets dropped.
-pub type LabelGuard<'a, L> = parking_lot::MappedRwLockReadGuard<'a, L>;
 
 struct TypeMap(StableHashMap<TypeId, Box<dyn std::any::Any + Send + Sync>>);
 
@@ -157,6 +203,7 @@ impl Labels {
         set.get_index(key as usize).map(f)
     }
 
+    /// Gets a reference to the label with specified index.
     pub fn get<L: 'static>(&self, key: u64) -> Option<LabelGuard<L>> {
         RwLockReadGuard::try_map(self.0.read(), |type_map| {
             type_map.get::<IndexSet<L>>()?.get_index(key as usize)